Remove variadic.go
* Since only variadic delete operations supported, move variadic logic inside `cmd_delete.go`. Also dont create dummy goroutine for copy operations, since only one source is needed.
